Improved protein filter settings text check
Former-commit-id: 13fb26713d576d6b111210af811dd2603cd5d3e1 [formerly b55e68370e46b1bc9ca8afab3948b1191ffc3a04] [formerly 629fdef0860797a74463b685bd443bcad6635d5f]
Former-commit-id: 5913c058cd60de63b302e425615fe2cf56daad78
Former-commit-id: 5b4c38017ae97879fcff3e8179916d601754c161

+++ TriFusion.py
@@ -5934,13 +5934,13 @@ if __name__ == "__main__":
 
             try:
                 # Check if max_stop value is between 0 and 100
-                if 0 < int(max_stop) > 100:
+                if 0 < float(max_stop) > 100:
                     return self.dialog_floatcheck(
                         "ERROR: Maximum codon stops value must be between 0 "
                         "and 100", t="error")
 
                 self.protein_min_len = abs(int(min_len))
-                self.protein_max_stop = abs(int(max_stop))
+                self.protein_max_stop = abs(int(float(max_stop)))
                 self.dismiss_popup()
             except ValueError:
                 return self.dialog_floatcheck(
